backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.12.0
djangorestframework-simplejwt>=5.0.0
psycopg2-binary>=2.9.0
django-cors-headers>=3.10.0
django-filter>=23.0
argon2-cffi>=21.3.0
python-decouple>=3.6